Allow running mistral-api service via wsgi in container
A review to add the necessary packages to run mistral-api via wsgi was
proposed to kolla in I07ffd0b398e20065fff27c30d0842b94090ca0ad but this
is unlikely we'll be able to backport to stable/pike due to the nature
of the change.

So make the necessary change in our override file so we have more
freedom to backport.

Change-Id: Id2bfb6ba761821616301f996dd68fea8f07ae5f3
Related-Bug: #1724607
(cherry picked from commit 1a4da389bb2e3d7bf703ebc0dd10fc46a99bd554)