Add work_dir for some actions
The Ansible actions and DownloadConfig actions both relied on a tmp path
for the working files they saved. This is less than ideal as some of
this data could be sensitive so it should not be stored in /tmp.

Also, the path was unpredictable, making things more difficult from a
user perspective.

This patch adds a work_dir input to these actions and makes use of that
input from the config-download-deploy workflow. The default behavior
when this input is not supplied is the previous behavior to use a
generated path under /tmp.

Change-Id: I89b17d0c4f035705fd5230caad06d331fd5b5e76
implements: blueprint ansible-config-download