Add rsyslog sidecar image for container image prepare
Some services (namely HAProxy and swift) use syslog for their logging,
and will need a sidecar container that runs rsyslog to redirect the logs
to stdout. So, when these services are enabled and we set the variable
"logging" to "stdout", we then enable the rsyslog container image as a
parameter.

The rsyslog-base image was used since we don't need any plugins for the
sidecar container.

bp logging-stdout-rsyslog

Change-Id: I1bb6db4a80c601790800a6d5657c9f114d0d8e74