Correct the failed status in the baremetal workflow
Two tasks in the introspect workflow have incorrectly been sending
"ERROR" statuses when they should have sent "FAILED". This then meant
the workflow appeared to finish without errors (or failures). This is
primarily a problem for the introspect_manageable_nodes workflow which
then can't detect errors and reports that everything was successful.

Change-Id: I34a91dd14bb19775ad62271def6ecb66398c84db
Closes-Bug: #1733303