Handle 'false' in when statements for ansible upgrade_tasks
The upgrade_task that remove packages have a when clause that is
set from the UpgradeRemoveUnusedPackages parameter. When we write
the overcloud config and translate "tags: step" to "when: step"
we must be careful to handle these tasks otherwise we get errors.
See the bug for more info.

Change-Id: Ic10ae0f3262a93ceec6ba8d938f36c937f3b594c
Closes-Bug: 1734118
(cherry picked from commit 2a9012a3f0ffcac7d946581b68f14daf118939c3)