Log the error from OrchestrationDeployAction
If the action fails, it returns an error to be used in the workflow.
While this is useful it means there is little trace of the error in the
log files. Adding the error also to the logs will make it easier to
debug and track down.

Related-Bug: #1734747
Change-Id: I5faba1feec7e87bb595edb3448a096caa0812a9b
(cherry picked from commit 5d4cadef9eff3e681e61b57c7bb4a0f975c52d1d)