Use push_destination as the registry host in env file
When using the undercloud registry, the prepare call needs to be run
twice, once to generate the upload instructions file, and again with a
different --namespace to generate the environment file.

This can be combined into a single call by modifying the env
paramaters by replacing the host portion of the registry path to use
push_destination. That is what this change does.

Change-Id: I7ad0ae8dee665c8c7d0599eb6d29787bf9043762
Closes-Bug: #1738663