Use skopeo for tag discover
Calling "overcloud container image tag discover" is slow because the
only docker library interface for inspecting an image requires pulling
the image locally first.

This change uses a single call to "skopeo inspect" to do the tag
discovery, which is significantly faster.

Since skopeo calls require an extra argument to disable TLS
verification, an is_insecure_registry implementation has been added
which includes memoization, and a whitelist of known secure
registries.

The implementation is split into static functions to help the
implementation of a concurrent tag discover which works on a list of
images.

Change-Id: I9c18f01d081bebb51246e1c65dc6d6b00f1c7cf6
Closes-Bug: #1738666