Allow separate parameters to TripleoInventory class
In the case of tripleoclient using this class there's no compatible
config object to pass, so it's easier and clearer if we break out the
individual parameters into the class constructor arguments.

The tripleo-validations code will be adjusted to suit then we can remove
the backwards compatibility code, but for now both old/new interfaces
are maintained.

Change-Id: I138ed6d07bd7ebb161b0a1dec3480de811e18e49