Add write_static_inventory to TripleoInventory class
This is a reimplementation of the similarly names function in
tripleo-ansible-inventory, but this outputs yaml format instead
of the current ini format.

The reason for this change is getting consistent behavior of the
configparser module for unit tests over py27 & py35 is difficult,
and also some of the roles_data keys I plan to add in a subsequent
commit can't easily be represented in the ini format (particularly
lists don't seem to work, everything must be a string value).

The tripleo-ansible-inventory script can be changed to use this
interface, and we'll have to deprecate the current ini based interface.

Change-Id: Ia3c6b9931e2d38507917cd80a93857d952784bd2