Remove step_tags_to_when from config download and fix sorting
We are now going to have 'when' in the upgrade_tasks directly
so no need to overwrite anymore (see depends-on).

This also changes the sorting of tasks and the order in which
they are written to the ansible playbook. We were previously
sorting on tags and this no longer works. Instead sort on the
step number given in the 'when:' value.

Closes-Bug: 1743758
Depends-On: I6adc5619a28099f4e241351b63377f1e96933810
Change-Id: I68c032c5f23393a77109f182541cc6bfcc019a41