Skip unnecessary image uploads
This modifies the container image upload to avoid doing an upload when
the destination image already exists and is identical to the source
image.

This is done by doing a skopeo inspect on the source and destination
images and compares their digest hash.

In RDO Cloud, a full upload takes around 8 minutes. With this change
and an already populated local registry it takes less than 10 seconds.
The extra 2 skopeo inspect calls per upload doesn't measurably impact
the full upload time.

This makes the upload fast enough to run within the main deploy
workflow on every run (which will take more work because the upload
yaml file needs to be stored in the plan).

Change-Id: Ic9422b0ca168b4405cb1973f4e33c91b0f69f02c
Closes-Bug: #1738668