Alter default for --min-tmpfs
This patch alters the default for the --min-tmpfs flag increasing the
chances of successful image builds in memory restricted environments.

Increasingly, and especially with increasing support for things like
OVB and other virtual deployments, people are deploying underclouds
in somewhat constrained environments.  A regular image build fails
routinely in an undercloud with <= 12GB RAM.  Unfortunately this is
tricky to nail down as the cause since typically the VM undercloud
locks up solid making debugging difficult or impossible.

Increasing this value means image builds complete successfully by
deferring the ramdisk meaning a happier experience for the operator
and negligible difference in the speed of the image creation.

Secondly, this patch adds consistency to flags by standardizing on
the = character.

Change-Id: I8fcd240b7063b4991404e01c08441fe929ff5e69