Format group:ansible deployments
The group:ansible deployments were not formatted as human readable in
the group_vars. It was all just one long line. This made manual review
and debugging more difficult. It should instead be in a human readable
format.

To accomplish that, we first attempt to convert the config value to
JSON, then pretty print it.

Change-Id: I1916d0378d4687966ec7b01f022e70c10e813f3c
Closes-Bug: #1745241