Make container image upload more resilient
This change implements a retry loop for the image push, just like the
existing one for pull. This will be necessary for some non-undercloud
registries which might periodically fail on some pushes.

This change also determines the upload worker count to be half of the
CPU count, with a minimum of two. This is to make CI more resilient on
smaller (6 core) flavors. A docker pull causes high CPU load in
dockerd, as the download of each layer is spread across dockerd's
workers.

Change-Id: Ia30658e3283d4b69d2bd8b0dddd375e1918169d3
Closes-Bug: #1746305