Add file type check to upload-swift-artifacts
deploy-artifacts supports only 'RPM' and 'gzip compressed data' [1]
while upload-swift-artifacts will allow the uploading
of any file. deploy-artifacts will fail if an unsupported file
type is downloaded resutling in a stack deploy failure.

This update add a check_file function which will validate the
file type of each file before it is uploaded. The script will
exit with a failure if an unsupported file is attempted.

Change-Id: I8c46110fa2779b818bc5b7882788f5d8739c1d2a
1: https://github.com/openstack/tripleo-heat-templates/blob/master/puppet/deploy-artifacts.sh#L12-L23