Provide the output to add a validation for role-specific params on services
In order to provide role-specific parameters for a service,
"<RoleName>Parameters" should be used to provide the actual
parameter to be applied to a specific role (in the parameter_defaults
of user's environment file). Parameters require special implementation
to support role-specific in a service. This workflow validates the
user provided role-specific parameters with the actual implementation
and provides the invalid ones. The actual role-specific parameters
on a template could be found using the 'tags' attribute of a
parameter in the heat template.

Change-Id: Ie0abc345cea4a69b19b2ea760a29a2cab61c1ceb
(cherry picked from commit 8586cc1542e894e52767507e1ec87f3d9ac95e03)