Use tenacity for image upload retries
This change replaces the custom image pull/push retry loop with a
tenacity annotation.

The docker client uses requests, so this fixes an issue where requests
can raise a vendored urllib3 exception if the registry doesn't respond
at all. This will now be caught by tenacity and treated as a retry
condition.

Change-Id: Ifa08f21ccbc1bf04732462d28449546cf146c53a
Closes-Bug: #1749663