Check mode support for Server Deployments
Adds a task for overcloud config-download server deployments that will run
only during check mode to test if the deployment file exists, and reports
whether ansible would make a change.

Makes the "Output for" and "Run deploment" tasks not run in check mode.
shell tasks are implicitly not run by ansible in check mode, and the
"Output for" task depends on the deployment_result var registered by
"Run deployment".

Change-Id: I7c310f52d872b4fd03e9bfa4a6e2e63684350f07
Closes-Bug: 1750438