Remove the noop deploystep for upgrade converge step
The noop deploystep is set during the prepare step for
upgrade. It put in the plan_env directly, so since we are
keeping the user-env / plan-env for upgrade the noop
steps are never removed.
This patch aim to trigger a mistral action to remove all
the OS::Heat::None for the deploystep/pre/post

Change-Id: If7eb342a84b530b07db513146e85916f2d5c1370
(cherry picked from commit 426f4aa6d859eda7614cd173974f5ccd32446116)