Add excludes option to create_tarball function
This option will allow you to pass in a configurable list of
directories/files that will be added onto the resulting tar
command with a --exclude option.

This is useful elsewhere in tripleo in that it will allow us to
further prune the growing t-h-t tree a bit to optimize performance a
bit.

Change-Id: I2be733efd199a503d4ccc90a01764576fb5fe78b