Retry previously failed deployments
Previously, when the server pre and post deployments were run, if one
failed, it would not be automatically retried since the deployed file
for the deployment exists under /var/lib/heat-config/deployed.

This could be overridden by passing -e force=true to ansible-playbook,
however for deployments that have previously failed, I don't think this
should be required, and we should just retry them automatically.

This new behavior more closely matches the traditional behavior with
Heat, where a stack update would automatically attempt to reapply and
deployment that was in a failed state.

Change-Id: I18df88297abe7be50364233ed8f2514de6d3d434
Closes-Bug: #1756921