Use the standard messaging in the derive_parameters workbook
Update the derive_parameters workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Related-Bug: #1757372
Change-Id: If9bcf6f0348d06041754522d30cb1a110ea7186d