Use the standard messaging in the plan_management workbook
Update the plan_management workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Related-Bug: #1757372
Change-Id: I9755eaa6621ec1d83fe98e20445a4f13b4ea0686