Raise an error if deployment_name cannot be determined
Parameter deployment_name should be set via the name property on
the Deployment resources in the templates, however, if it's None,
it will be defaulted to the name of the parent_resource.
for resources such as ExtraConfigPost, PreConfig, PostConfig,
there is no default parent resource name which means the
generated ansible group_vars will be an empty string if custom
templates don't contain a name property; this in turn will
cause syntax error when executing ansible playbooks.

This code prevents such situations and raises an error if
the value cannpt be determined.

Change-Id: Ibe884331b5fb9269b796a13beb489e91aeca71ed
Closes-bug: #1755078
(cherry picked from commit 3fd4f24e90a2686f20472a78d59b4ce1963cdb62)