Represent image prepare defaults in new heat param format
The heat parameter ContainerImagePrepare is a list of dicts where each
entry is a single prepare call. This represents the defaults in the
new format in preparation for a function which consumes this new
format.

It removes tag_from_label from the substitution dict, because its not
a substitution, its an actual argument to prepare.

Change-Id: I8952ce7a19a67b89d8db5e90fbe65620425ee945
Blueprint: container-prepare-workflow