Add openvswitch element back in
The change Ie4f5d771a16ea453b470be8ea103b2bde4aa892a switched
os-net-config to be installed as a package rather than an element.
Unfortunately the openvswitch service was being configured via the
openvswitch dependency from the os-net-config element. The switch to use
the package lost the automatic starting of openvswitch on the nodes
which has manifested itself as intermittent issues on compute nodes when
network isolation is not deployed.

Change-Id: I0f865d4811919c19577c75615b66d7d8a1e685d3
Partial-Bug: #1757111