Remove /etc/os-net-config/config.json in os-apply-config
In the new style of nic config files, os-net-config is invoked by
the script run-os-net-config.sh, not by os-apply-config.  This is the
only format supported in Queens. Remove the os-apply-config version
of /etc/os-net-config/config.json to prevent os-apply-config from
attempting to run with an empty config.json file.

Note that /etc/os-net-config/config.json no longer exists in
tripleo-image-elements.

Change-Id: Ica86382c4e5a1463ed8b2146e3ec490b7e2c9452
Closes-Bug: 1758161