Enrich integration with ironic networking features
The mac key in nodes_json is deprecated, replaced with
"ports" key. New ports key is list of dicts holding a
richer data set matching the properties of ports in the
Bare Metal service api. In addition to mac address the
physical_network and local_link_connection can be
defined for Bare Metal ports when registering nodes.

* address: (mandatory)
  The physical address (mac address) of the port.
* physical_network: (otional)
  Defaults to: ctlplane
* local_link_connection: (optional)
  This data enables the possibility for automatic
  configuration of switches via neutron plugins. e.g
  ML2 vendor plugins. Defaults to: None

Implements: enrich-ironic-networking-integration
Change-Id: I74d4178dbb0cfe8c934ce15e3e7c9bb1c469de10