Remove the upgrade-non-controller.sh and disable_upgrade_deployment
The disable_upgrade_deployment flag is removed from the roles_data
(see depends-on) and so we remove it from the roles schema. The
upgrade-non-controller.sh is also no longer used (last used in
Pike) and so removed. The two are related as they were part of
the same upgrade mechanism.

Depends-On: Ib340376ee80ea42a732a51d0c195b048ca0440ac
Depends-On: https://review.rdoproject.org/r/#/c/13421/
Change-Id: I7b19c5299d6d60a96a73cafaf0d7103c3bd7939d