Use scheduler hints in derived_parameters workflow
Update the tripleo.derive_params.v1.derived_parameters workflow to use
scheduler hints associated with a given role. The scheduler hints are
used to identify overcloud nodes associated with the role, and take
precedence over nodes identified by their profile/flavor. If a role does
not specify any scheduler hints then the workflow falls back to matching
nodes by profile/flavor.

Add a tripleo.baremetal.v1.nodes_with_node_id workflow that searches
for ironic nodes matching a node placement regex. This workflow is
analogous to the existing nodes_with_profile workflow that searches
for nodes that match a given profile.

Add a tripleo.baremetal.get_node_id Mistral action that returns the node
placement ID (e.g. "controller-0") for a given node, or None if a node
ID is not defined. This is analogous to the existing get_profile action
that returns the profile assigned to the node.

Fix the 'message' published when a derived parameter sub-workflow fails
so it just returns the sub-workflow 'message' and not the entire sub-
workflow result. The 'message' portion is the only part that is meaningful
to the user.

Closes-Bug: #1760659
Change-Id: I7eff355620aecaca49e77112ba491a5f3ce2eed6