Add health checks for Swift services
This patch is adding health checks for background processes. The check
is performed by comparing recon cache file modification. This is far
from ideal, because recon files are shared via services of same type.
But it's the only way to at least partialy check correct behaviour
for background processes. In future there we should add timestamp info
to each recon file for all services using the recon file. Then we would
be able to find out if all background services are updating their caches
in timely fashion and stay within the border of containers.

Co-Authored-By: Thiago da Silva <thiago@redhat.com>

Change-Id: Ib6fad8311b5a728914ce9df9122194c5f7036be7