Allow for more overhead in zaqar message
We need to account for more overhad when chunking the message before
posting it to the zaqar queue as JSON encoding also adds to the overall
message length. This changes the overhead ratio from 90% down to 50%.

This would result in more smaller messages going through the queue. I
don't see any drawbacks with that though given this zaqar is not tenant
facing and is on the undercloud.

Change-Id: I79535bb276a7f763da575a022486590c117579b0
(cherry picked from commit dec6aae5116c4e7f86e6796b74d4c75e5ed8ce43)