Find deployment name in config download
In some cases we get a ResourceGroup as parent resource of a deployment,
so the name will be a number. Ansible will be unhappy with that, so
let's get the grand parent to get the proper name.

Related-Bug: #1755078
(cherry picked from commit 635f6a4374b35489084131d62b88e7272b987bb3)
Change-Id: Icc64b32a2179cd10853a54a83bf82a1761297ebd