Add in a STOPSIGNAL configuration
For docker 1.13.1, if STOPSIGNAL is not provided in the docker image,
the container will not automatically restart the container.  It might be
better to define this in the Kolla scripts as a configuration option for
each container, but I'm unsure if it's a problem for any other version
so it might be best to configure this in tripleo until a newer version
of docker is consumed. Additionally I'm not sure if any of the
containers would need something other than STOPSIGNAL SIGTERM which
should be the default if none is provided.

Change-Id: I1939f9e6b2c432a672c7426ddabdcfca6ce150b7
Related-Bug: #1765802
(cherry picked from commit 039dbccb5d568953e281a21854916b60ad0a98e4)