Fix missing docker package in neutron L3 agent container
We need to use the append macro to add the docker package to the L3 agent
container.

Note: the dhcp image should really be doing the same thing, but requires
a change to the kolla repo and will be addressed in follow-up patches.
(See https://bugs.launchpad.net/kolla/+bug/1766863)

Change-Id: Ib2d2ad4960ea34ec9e3fca1eeb322742341f7eb7
Related-Bug: #1749209