Add workflow for plan deployment status
Add a new workflow used for checking the plan deployment status. With
the move to config-download, the Heat API is no longer the source of
truth for deployment status.

This patch makes use of the top level deployment_status added in
Ib854689a3e606f18db7354427918fa8043ac28e0 to expose the deployment
status via a workflow.

Implements: blueprint config-download-default

Change-Id: Ia8e3ecadfb97c0827c396cd418c32d3daa2e299b