Script to update RPM's in containers.
This script was living in a github repo and is used during the CI
job.  Moving it here so more people can review it.

This updates the RPM's in a container without having to rebuild
the whole thing.

Change-Id: I4905c35fcd45ef0fe1845857020c7e1a44c8c3d6
(cherry picked from commit deb8480163254edf6497ad6149be25c94d1db36f)