Add a 'trash_output' flag to ansible playbook action
Sometimes, the output is REALLY big. It ends up throwing an error in
MySQL trying to write to the output of the ActionExecution to the
database. Since we're passing the output back via zaqar, we can
just throw this away.

The trash_output flag defaults to false, to maintain the original
behavior.

Change-Id: Id13c7cf52b306891d780243b2562c8ca305aa365
Closes-Bug: #1768913
(cherry picked from commit 89986793a8c3bff9be28cc674923b9f30eab1e89)