Allow image prepare to only filter by role data
Currently if no resource registry is in the environment passed to
prepare, then every image will be filtered out of the result.

Sometimes it is useful to only provide the role data and still have
filtered images on what services are enabled in the roles. This change
makes that possible

This change also returns a None service filter when provided None
roles data. This is to ensure no filtering occurs when no role data
file is specified.

Change-Id: Id2b787f148709506df4a0619103850051f786e29
Closes-Bug: #1769549