Install Octavia amphora image if Red Hat
For a one-click deployment, the Octavia amphora image must be present in
the undercloud node. Currently users need to ensure the image file
exists before overcloud deployment. For Red Hat-based deployments, the
image is provided by package octavia-amphora-image. Thus the package
should be installed for a better user experience leading to a truly
one-click deployment.

Closes-Bug: #1770117
Change-Id: I3ed8144b52ba528f449a4abe5c2106a5c61f9d61