Check pub key file perms and default to pub key data
The previously default /home/stack/.ssh/id_rsa.pub file may not exist or
be readable; exit with explicit error message. Users can still specify a
file path but will need to ensure it is readable. Should a file path not
be specified, default to amp_ssh_key_data. The value is passed by THT
with the public key of the 'default' keypair from the undercloud which
anyway is the public key of the 'stack' user.

This patch also fixes a syntax error in octavia-undercloud role.

Closes-Bug: #1770153
Closes-Bug: #1770641

Depends-On: https://review.openstack.org/568022
Change-Id: I0026343d90b84572c3002fa21001cfb09c742391