Make lb-mgmt-subnet a class B subnet
TripleO creates that subnet during the Octavia deployment process.
Currently, it is created as a class C subnet with allocation_pools
that essentially limit the number of address in that subnet to 150.

This patch enhances the lb-mgmt-subnet to be class B so the global
amount of Octavia loadbalancers won't be constrained to a very low
number.

Closes-Bug: #1770956

Depends-On: I3465dd9e2f95983a3f3a5dcd85f50781044fd6ec
Change-Id: I41b98f7782f055533bf7dd61e2d73ded2829c895