Add validation for empty file system backup on Undercloud
This patch works as an addition to the exclude path
when executing an Undercloud backup.

In this case if we exclude all files, there is no need to execute
the task to create the FS backup.

Change-Id: Icd25a8d579151015c9cdf56861d2be933848a723
Closes-Bug: 1767385
(cherry picked from commit 650a259472de77c1ea6ecd6883b50c134cd6d29a)