Add yum update to base
In order to handle out of date distro containers,
let's add a yum update to the kolla base container to ensure that
at the end of the base container build we'll have the latest base
packages installed.
This will ensure we don't hit out of date dependency bits in later
containers.
For example, we currently only have 7.4 base centos image available
from the upstream but we're install the base repos in the base
build. This would mean that we'd be getting 7.5 packages in later
containers but might possibly have a mix of 7.4 and 7.5 packages
installed.

Change-Id: Ifdfc1b84b7b6a306868260611271fd49bdd594d4
Related-Bug: #1770355 (cherry picked from commit 3f12a00533e32d894ad18627e6be1163c0dd5ad8)