Ignore errors when checking result of previous deployments
This always should have been ignore_errors:true, as we don't want to
fail when checking the result of the previous deployment. Instead, we
want to use the result as a trigger as to whether or not we should
re-run the deployment.

Change-Id: I6b75ea2a6096e152d7f01853de5da59cccc8e9ec
(cherry picked from commit 4dd9a779eb6555fa1e981f2aba431c3475b151f0)