Clean up node registration output.
The cell_v2_discover_hosts is an internal workflow
and doesn't need to send messages. It is retried by
the parent workflow as it is expected to fail, and
when it does we don't want it to send error messages
each time. which can be retried up to 30 times by
the parent workflow.

Change-Id: I4b8f8d153a8e8c2eabb9efadb52cf685e1f3759f
Closes-Bug: #1710685