Add a dry-run behaviour to image image_uploader
It will sometimes be required to generate the heat environment
containing the final image parameters without doing any
pull/modify/push operations (such as when the push registry hasn't
been installed yet).

This change also allows the append tag to be passed in with each
ContainerImagePrepare via the modify_append_tag option, so that the
same append tag can be used for the dry-run and the real prepare.

Change-Id: I9708b151a556f0517df8cbf8492e68ef1064745d
Blueprint: container-prepare-workflow