Test target image exists before modifying
When an image is simply uploaded, the source and target image digests
are compared so upload can be skipped if they are the same.

When an image is modified by an ansible role before upload, the digest
of the target will always change, so this mechanism can't be used to
skip unrequired modifications.

With this commit, if there is a modify role set, and the target image
already exists, then the pull/modify/push will be skipped.

This allows the append tag to drive whether the modification happens
for an existing image. Set it to a different timestamp on every
invocation to force the modify to occur.

The intent is to use the same append tag for a single quickstart run,
so that overcloud modify will be skipped for images already modified
during the undercloud deploy, saving significant time during the
overall deployment.

Change-Id: I7a4dc0625c212f6361793609a3558bba248b151f
Blueprint: container-prepare-workflow