Increase services project secgroup-rules quotas when deploying Octavia
This patch is a followup to I6b87b147d301dea3251fe4509b04cd4b9b27ddba,
which lacked the secgroup-rules quota.

The reasoning is the same:
Octavia currently launches load balancer VMs in the services tenant
which has very low quotas by default. This patch increases the values
through the API for the services project to permit lots of load
balancers.

Closes Bug: #1769896

Change-Id: I8ad5175f7ca79939725a0c2342378a052b5c6bd2