Remove container_registry param from package update action
The package_update_plan mistrl workflow is invoked in upgrade|update|
ceph-upgrade|ffwd-upgrade prepare clis. The container registry
parameter is being removed and we will use treat it as a regular
environment file (see depends-on). This removes it from the Mistral
action and keeps it defaulted in the workflow to prevent
chicken-and-egg issues with CI. We'll remove the workflow parameter in
a follow-up patch.

Partial-Bug: #1785825
Co-Authored-By: Jiri Stransky <jistr@jistr.com>
Related: https://review.openstack.org/570893 python-tripleoclient
Change-Id: I4130dd1eea6caec82dcb4a699875f2951693148f