Deploy ceph-ansible in mistral containers
It's required when containering the undercloud and we deploy Ceph on the
overcloud, we need the module in the container to execute the Ansible
tasks.
The dependency has to live here and not in the distgit because we need
the ceph repos in place, which is already managed here when building
containers.

Change-Id: I5dbb12adddeedbb92ba9963f31e2dd822fbb6cce