Use the standard messaging in the swift_rings_backup workbook
Update the swift_rings_backup workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Change-Id: Iad199545d0093ccdedf082a945345ad4863ed4d0
Related-Bug: #1757372