Don't return image tags during dry_run=True
If images paths are returned during dry_run upload, the cleanup will
be triggered on those images and an attempt will be made to delete
them from the local docker.

However one use-case for dry-run is when there is not yet a configured
local docker service, so this will fail.

This change avoids creating a docker client when there is nothing to
cleanup, and avoids returning images which are not actually tagged in
the local docker.

Change-Id: I8aca9aaf42fe92e4fbea4aa93076900101091c40
Blueprint: container-prepare-workflow