kolla_builder: use any templates except puppet
We have more and more services that use containers now, and not only
OpenShift, k8s. We now have Skydive and probably more services in
extraconfig/services that will come in the future.

This patch will use all but puppet templates to determines where are the
containers that we want to prepare with kolla.

Change-Id: I16a9e8533ba8cf1553ccac8c43f7556a71bd5b15