Provide timestamps in Ansible output by default
Add a new configurable `profile_tasks` argument for
AnsiblePlaybookAction that controls output of profiling
information. The default for the new option is True, as it's likely
that this will be a universally useful default for deployment,
updates, upgrades, etc.

Also a `profile_tasks_limit` argument is added to control final output
of slowest executing tasks in the playbook. Since we want to optimize
for production scenarios and operators can hardly affect the run
times, this argument defaults to 0.

The combination of the above will give us time stamps for each task
but no "what are the slowest tasks" by default.

Both options are overridable by extra_env_variables, which might be
useful in some cases, especially for the callback whitelist.

Change-Id: If8abc8c2406e152004c7bce1d116bc8690ac1cf7
Closes-Bug: #1777467