Use the standard messaging in the package_update workbook
Update the package_update workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Change-Id: Icad2846b43dec8032d9692daea390bffa7486dac
Related-Bug: #1757372