Use the standard messaging in the plan_management workbook
Update the plan_management workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

This is a continuation of I9755eaa6621ec1d83fe98e20445a4f13b4ea0686.
Some old-style messaging was added after the refactor and some redundant
lines were not removed.

Change-Id: Ia50e4055e24d07cb39d4c05745bd6694e94621f9
Related-Bug: #1757372