Use the standard messaging in the parameters workbook
Update the parameters workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Change-Id: I010b848721732a7a186dd4de183089b5e24bfd04
Related-Bug: #1757372