Use the standard messaging in the scale workbook
Update the scale workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Change-Id: I906e0ae92bcd084f4bf6ec669d6be0f71dc7a29b
Related-Bug: #1757372