Use the standard messaging in the undercloud_backup workbook
Update the undercloud_backup workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Change-Id: I71e761c1295ae0d0ca2773c7521b6a57cabd6227
Related-Bug: #1757372