Use the standard messaging in the stack workbook
Update the stack workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

Change-Id: I2bf3b0eea591969892d8c98937f62719c4e02361
Related-Bug: #1757372