Use the standard messaging in the baremetal workbook
Update the baremetal workbook to use the standard message sending
workflow. This workflow reduces the boilerplate that is copied between
workflows by defining the message structure, providing the retry logic
and ensuring the task on a "FAILED" message.

The workflow also offers optional persistence in Swift. If the plan name
is given to the workflow it will store the messages in a Swift container
named <plan_name>-messages.

This is a continuation of I1665ae56a88557bdb8459651021185cfb9fd1a8f.
Some old-style messaging was added after the refactor and some redundant
lines were not removed.

Change-Id: Ieeabe23ffe134cf19634b4b052805ae7165f34e6
Related-Bug: #1757372