Use upstream etcd container image for openshift
The kolla based etcd image doesn't work with openshift-ansible [1]
because it requires to read from the bind-mounted /var/lib/etcd
directory and there is a uid/gid mismatch with the etcd user in the
kolla-based container. The upstream etcd container doesn't have this
issue because it runs as root. Unfortunately there is no way currently
in openshift-ansible to set the etcd container user at runtime [2];
we're stuck with an etcd image that runs as root by default.

[1] http://paste.openstack.org/show/723815/
[2] https://github.com/openshift/openshift-ansible/blob/master/roles/etcd/templates/etcd.docker.service#L10

Depends-On: I81e5cd0edae56c1535a2e4dcc639ec71055b10d2
Change-Id: I3697c171243ec2f3001cca83f7a93ece2af9a8bd