Switch ansible tmp for local connections
With ansible 2.5.4, they introduced a change that causes the temp
directory used to not honor the HOME for a user. See
https://github.com/ansible/ansible/pull/40363. Now the directory
defaults to ~/.ansible/tmp but using the HOME env has no effect. In our
case the code can be executed via mistral which doesn't have a valid
or writable home directory so it errors.  This change updates the
ansible_remote_tmp for the undercloud inventory file to
/tmp/ansible-${USER} to ensure it will work when running under mistral
but will work for any user as well.

The recommended fix for this is to configure the ansible_remote_tmp in
the inventory file for local connections to change this location.
Alternatively ANSIBLE_REMOTE_TMP can be used via an environment var but
that would change the location for all hosts which is not ideal.

Additionally HOME has been replaced with ANSIBLE_LOCAL_TEMP due to
changes upstream around the effect of home. See
https://github.com/ansible/ansible/pull/36755

Change-Id: Iaf1b8aa68af0274ba40eae329e32ab14f3689481
Closes-Bug: #1778269
(cherry picked from commit f7bd4e71fe092724780de37002275016a29e5f3d)